requirements.txt
# Core dependencies
discord.py>=2.3.0
anthropic>=0.5.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.5.0

# Database and caching
aiosqlite>=0.19.0
redis>=5.0.0

# Task scheduling
apscheduler>=3.10.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Optional dependencies
html2text>=2020.1.16  # For message content cleaning
aiohttp>=3.9.0        # For webhook delivery